# backend/swagger/swaggerconfig.yaml
openapi: 3.0.0
info:
  title: For me
  description: 설문지 폼 커스터마이징할 수 있게 해주는 API
  version: 0.1.9
servers:
  - url: http://localhost:8000
paths:
  /api/surveys:
    post:
      summary: 설문 조사와 이미지를 업로드합니다.
      tags:
        - POST
      description: 메인 이미지와 질문 이미지를 포함한 설문 데이터를 업로드합니다.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                survey:
                  type: string
                  description: JSON 형식의 설문 데이터 문자열.
                mainImageUrl:
                  type: string
                  format: binary
                  description: 메인 이미지 파일.
                imageUrl:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: 질문 이미지 파일 배열.
      responses:
        '201':
          description: 설문 생성이 완료되었습니다.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 설문 생성 완료
                  surveyId:
                    type: integer
                    example: 9
        '400':
          description: 요청이 올바르지 않아 설문 생성에 실패했습니다.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid survey
                  error:
                    type: string
                    example: No survey data provided
        '500':
          description: 서버 에러로 인해 설문 생성에 실패했습니다.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 설문 생성 오류
                  error:
                    type: string
                    example: Internal server error

  /api/surveys/{userId}/answers/{surveyId}:
    get:
      summary: 내 응답
      tags:
        - GET
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: The unique ID of the user.
        - in: path
          name: surveyId
          required: true
          schema:
            type: integer
          description: The unique ID of the survey.
      responses:
        '200':
          description: A user's answers to a survey
          content:
            application/json:
              schema:
                type: object
                properties:
                  mainImageUrl:
                    type: string
                  title:
                    type: string
                  description:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
                  deadline:
                    type: string
                    format: date-time
                  userId:
                    type: integer
                  questions:
                    type: array
                    items:
                      type: object
                      properties:
                        questionId:
                          type: integer
                        content:
                          type: string
                        imageUrl:
                          type: string
                        subContent:
                          type: string
                        objContent:
                          type: array
                          items:
                            type: string
        '400':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User not found'
        '404':
          description: User has not responded to this survey
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '사용자가 이 설문에 응답하지 않았습니다.'
        '500':
          description: Failed to retrieve user's answers
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '사용자의 답변을 불러오는 것을 실패하였습니다.'

  /api/surveys/{id}/join:
    get:
      summary: 내 응답 조회
      tags:
        - GET
      operationId: 'showAnsweredSurveys'
      parameters:
        - name: 'id'
          in: 'path'
          description: 'ID of the user'
          required: true
          schema:
            type: 'integer'
        - name: 'page'
          in: 'query'
          description: 'Page number'
          required: false
          schema:
            type: 'integer'
        - name: 'limit'
          in: 'query'
          description: 'Number of answered surveys to return per page'
          required: false
          schema:
            type: 'integer'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: 'object'
                properties:
                  surveys:
                    type: 'array'
                    items:
                      $ref: '#/components/schemas/Survey'
                  totalPages:
                    type: 'integer'
        '404':
          description: 'No answered surveys found for this user'
        '500':
          description: 'Server error'

  /api/surveys/{id}/all:
    get:
      summary: 공개 설문 조회
      tags:
        - GET
      operationId: 'showAllSurveys'
      parameters:
        - name: 'id'
          in: 'path'
          description: 'ID of the user'
          required: true
          schema:
            type: 'integer'
        - name: 'page'
          in: 'query'
          description: 'Page number'
          required: false
          schema:
            type: 'integer'
        - name: 'limit'
          in: 'query'
          description: 'Number of surveys to return per page'
          required: false
          schema:
            type: 'integer'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: 'object'
                properties:
                  surveys:
                    type: 'array'
                    items:
                      $ref: '#/components/schemas/Survey'
                  totalPages:
                    type: 'integer'
        '404':
          description: 'No surveys found'
        '500':
          description: 'Server error'
          content:
            application/json:
              schema:
                type: 'array'
                items:
                  $ref: '#/components/schemas/Survey'

  /api/surveys/{id}/share:
    post:
      summary: 설문 url 공유
      tags:
        - POST
      description: 특정 설문조사의 링크를 이메일로 공유합니다.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: 설문조사 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                emails:
                  type: array
                  items:
                    type: string
                    format: email
                  description: 공유할 이메일 주소 목록

      responses:
        '200':
          description: 이메일을 성공적으로 보냄
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '이메일 발송 요청 완료'
        '400':
          description: 요청 데이터 오류
        '404':
          description: 설문조사를 찾을 수 없음
        '500':
          description: 서버 오류

  /api/surveys/{id}/urls:
    get:
      tags:
        - GET
      summary: 설문 url 받기
      description: 특정 설문조사(survey)의 URL을 가져옵니다.
      parameters:
        - in: path
          name: id
          required: true
          description: serveyId
          schema:
            type: integer
      responses:
        '200':
          description: 성공적으로 URL을 가져왔을 때의 응답
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    example: 'www.asdasdasdasdasd.com/asddasd'
        '404':
          description: 설문조사를 찾을 수 없을 때의 응답
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '설문이 존재하지 않습니다.'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'URL을 생성하는데 실패하였습니다'
  /api/surveys/{id}/list:
    get:
      tags:
        - GET
      summary: 설문지 응답 결과 분석 폼(응답별)
      parameters:
        - in: path
          name: id
          required: true
          description: Unique identifier of the survey
          schema:
            type: integer
      responses:
        '200':
          description: A JSON object containing the survey results
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                    description: The title of the survey
                  list:
                    type: object
                    properties:
                      head:
                        type: array
                        items:
                          type: string
                        description: Array of survey questions
                      rows:
                        type: array
                        items:
                          type: object
                          properties:
                            userId:
                              type: integer
                            createdAt:
                              type: string
                              format: date-time
                            responses:
                              type: array
                              items:
                                type: string
                        description: Array of user responses
        '400':
          description: Survey not found
        '500':
          description: Internal Server Error

  /api/surveys/{id}/results:
    get:
      summary: 설문지 응답 결과 분석폼(질문별)
      tags:
        - GET
      operationId: 'getSurveyResult'
      parameters:
        - name: 'id'
          in: 'path'
          description: 'ID of the survey to fetch'
          required: true
          schema:
            type: 'integer'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Survey'
        '404':
          description: 'Survey not found'
        '500':
          description: 'Server error'

  /api/surveys/{id}/forms:
    get:
      summary: 폼 모아보기 및 제목 검색
      tags:
        - GET
      description: 특정 사용자에 대한 설문조사 목록을 페이징하여 가져옵니다. title 파라미터가 있으면 제목 검색을 수행합니다.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: 사용자 ID
        - in: query
          name: title
          schema:
            type: string
          description: 검색할 설문조사의 제목 (없으면 전체 목록)
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: 페이지 번호
        - in: query
          name: limit
          schema:
            type: integer
            default: 9
          description: 한 페이지당 항목 수
      responses:
        '200':
          description: 성공적으로 설문조사 목록을 가져옴
          content:
            application/json:
              schema:
                type: object
                properties:
                  surveys:
                    type: array
                    items:
                      $ref: '#/components/schemas/Survey'
                  totalPages:
                    type: integer
        '400':
          description: 필요한 파라미터가 없을 때 반환합니다.
        '404':
          description: 검색 결과가 없을 때 반환합니다.
        '500':
          description: 서버 에러 발생

  paths:
  /api/surveys/{id}:
    put:
      summary: 설문 조사 데이터와 이미지 수정
      tags:
        - PUT
      description: 수정된 설문 데이터와 이미지를 업로드합니다.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: 수정할 설문의 ID
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                survey:
                  type: string
                  description: JSON 문자열 형식의 설문 데이터
                mainImageUrl:
                  type: string
                  format: binary
                  description: 메인 이미지 파일
                imageUrl:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: 질문 이미지 파일 배열
      responses:
        '200':
          description: 설문 수정 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 설문이 성공적으로 수정되었습니다
                  surveyId:
                    type: integer
                    example: 9
        '400':
          description: 잘못된 요청으로 설문 수정 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 잘못된 설문 데이터
                  error:
                    type: string
                    example: 설문 데이터가 제공되지 않았거나 잘못된 질문 유형
        '403':
          description: 접근 또는 수정이 금지됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 무단 사용자 또는 설문이 잠겨 있거나 설문에 답변이 있습니다
        '404':
          description: 요청한 설문이 존재하지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 요청한 설문이 존재하지 않습니다
        '500':
          description: 서버 오류로 설문 수정 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 설문 수정 오류
                  error:
                    type: string
                    example: 내부 서버 오류

    post:
      tags:
        - POST
      summary: 응답 제출 버튼
      parameters:
        - in: path
          name: id
          required: true
          description: 설문 ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  description: 사용자 ID
                questions:
                  type: array
                  items:
                    type: object
                    properties:
                      questionId:
                        type: integer
                      objContent:
                        type: array
                        items:
                          type: integer
                      subContent:
                        type: string
      responses:
        '201':
          description: 답변 저장 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: 성공 메시지
                    example: '저장되었습니다.'
                  surveyId:
                    type: integer
                    description: 생성된 설문조사의 ID
        '400':
          description: 잘못된 데이터 입력
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '데이터를 찾을 수 없습니다.'

        '404':
          description: 잘못된 데이터 입력
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '설문에 해당 질문이 없습니다.'
        '409':
          description: 중복된 데이터 입력
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '하나만 선택해주세요.'

        '500':
          description: 요청 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '저장을 실패하였습니다.'
    get:
      summary: 설문지 응답 폼
      tags:
        - GET
      description: 특정 설문조사의 상세 정보를 가져옵니다.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: 조회할 설문조사의 ID
      responses:
        '200':
          description: 성공적으로 설문조사 정보를 가져옴
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Survey'
        '404':
          description: 설문조사를 찾을 수 없을 때의 응답
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '설문조사가 존재하지 않습니다.'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '서버 오류 발생'
    delete:
      summary: 설문지 삭제
      tags:
        - DELETE
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: 삭제할 설문의 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  description: The unique ID of the user.
                surveyId:
                  type: integer
                  description: The unique ID of the survey.
      responses:
        '200':
          description: 설문 및 관련 데이터 삭제 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: 성공 메시지
        '403':
          description: 설문 삭제 권한이 없습니다.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: '설문 삭제 권한이 없습니다.'
        '404':
          description: 삭제할 설문이 없습니다.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: '삭제할 설문이 없습니다.'
        '500':
          description: 설문 삭제 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: '설문 삭제 오류'
  /api/users/signup:
    post:
      summary: 회원가입
      tags:
        - USER
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: '회원가입 성공'
        '400':
          description: '회원가입 실패'
  /api/users/login:
    post:
      summary: 로그인
      tags:
        - USER
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: '로그인 성공'
        '401':
          description: '비밀번호가 틀렸습니다.'
        '404':
          description: '존재하지 않는 메일입니다.'
  /api/users/{email}/check-email:
    post:
      summary: 이메일 중복 확인
      tags:
        - USER
      parameters:
        - name: 'email'
          in: 'path'
          required: true
          schema:
            type: 'string'
      responses:
        '200':
          description: '중복 확인 결과 반환'
  /api/users/:
    patch:
      summary: 비밀번호 수정
      tags:
        - USER
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: '수정이 완료되었습니다.'
        '400':
          description: '비밀번호 수정에 실패하였습니다.'
  /api/users/{id}:
    get:
      summary: 내 정보 조회하기
      tags:
        - USER
      parameters:
        - name: 'id'
          in: 'path'
          required: true
          schema:
            type: 'integer'
      responses:
        '200':
          description: '사용자 정보 반환'

  /api/surveys/downloadExcel/{surveyId}:
    get:
      summary: 설문 응답 데이터를 Excel 파일로 다운로드합니다.
      tags:
        - GET
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: integer
          description: 다운로드할 설문의 ID입니다.
      responses:
        '200':
          description: Excel 파일 다운로드 성공
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
                description: 다운로드된 Excel 파일
        '404':
          description: 설문이나 관련된 질문이 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 설문이나 관련된 질문이 없습니다.
        '500':
          description: 서버 오류로 Excel 파일 생성 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Excel 파일 생성 오류

  /api/images/search/{query}:
    get:
      summary: 검색어에 따른 이미지 검색
      description: Pexels에서 제공하는 이미지를 검색어에 따라 조회합니다.
      parameters:
        - in: path
          name: query
          required: true
          schema:
            type: string
          description: 검색할 키워드
        - in: query
          name: perPage
          schema:
            type: string
            default: '6'
          description: 페이지 당 이미지 수 설정
      responses:
        '200':
          description: 성공적으로 이미지를 조회함
          content:
            application/json:
              schema:
                type: object
                properties:
                  photos:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: 이미지 ID
                        original:
                          type: string
                          description: 원본 이미지 URL
        '500':
          description: 서버 에러 발생
components:
  schemas:
    Survey:
      type: 'object'
      required:
        - id
        - userName
        - title
        - open
        - createdAt
      properties:
        id:
          type: 'integer'
          description: 설문조사 ID
        userName:
          type: 'string'
          description: '사용자의 이름'
        title:
          type: 'string'
          description: '설문조사의 제목'
        open:
          type: 'boolean'
          description: '설문조사의 공개 여부'
        createdAt:
          type: 'string'
          format: 'date-time'
          description: '생성된 날짜와 시간'
        updatedAt:
          type: 'string'
          format: 'date-time'
          description: '업데이트된 날짜와 시간'
        deadline:
          type: 'string'
          format: 'date-time'
          description: '마감일'
        questions:
          type: 'array'
          items:
            $ref: '#/components/schemas/Question'
    Question:
      type: 'object'
      required:
        - questionId
        - type
        - content
      properties:
        questionId:
          type: 'integer'
          description: '질문의 고유 식별자'
        type:
          type: 'string'
          description: '질문의 유형 (예: 객관식, 주관식 등)'
        content:
          type: 'string'
          description: '질문 내용'
        imageUrl:
          type: 'string'
          description: '질문과 관련된 이미지 URL'
        choices:
          type: 'array'
          items:
            $ref: '#/components/schemas/Choice'
        answers:
          type: 'array'
          items:
            $ref: '#/components/schemas/Answer'
    Choice:
      type: 'object'
      required:
        - choiceId
        - option
      properties:
        choiceId:
          type: 'integer'
          description: '선택지의 고유 식별자'
        option:
          type: 'string'
          description: '선택지 내용'
        count:
          type: 'integer'
          description: '해당 선택지를 선택한 응답자 수'
    Answer:
      type: 'object'
      required:
        - answerId
        - content
      properties:
        answerId:
          type: 'integer'
          description: '답변의 고유 식별자'
        content:
          type: 'string'
          description: '사용자의 답변 내용'
        objContent:
          type: 'string'
          description: '객체 형태의 답변 내용, 선택지 ID를 참조할 수 있음'
    User:
      type: 'object'
      properties:
        id:
          type: 'integer'
        email:
          type: 'string'
        password:
          type: 'string'
        name:
          type: 'string'